Move query-related stuff from types to 'query'
