+++ src/js/common/api/kontext/posQuery.ts
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { QueryPoS } from '../../types';
+import { QueryPoS } from '../../query';
 
 export interface PosQueryExport {
     (pos:QueryPoS):string;
