+++ src/js/common/models/concordance.ts
@@ -19,7 +19,7 @@ import * as Immutable from 'immutable';
 
 import { QuerySelector} from '../api/kontext/concordance';
 import { ViewMode } from '../api/abstract/concordance';
-import { LemmaVariant } from '../types';
+import { LemmaVariant } from '../query';
 
 
 export interface ConcordanceMinState {
