+++ src/js/tiles/concordance/index.ts
@@ -22,7 +22,8 @@ import { AppServices } from '../../appServices';
 import { QuerySelector } from '../../common/api/kontext/concordance';
 import { Line, ViewMode } from '../../common/api/abstract/concordance';
 
-import { DataApi, QueryType } from '../../common/types';
+import { DataApi } from '../../common/types';
+import { QueryType } from '../../common/query';
 import { Backlink, CorpSrchTileConf, ITileProvider, TileComponent, TileFactory, SourceInfoComponent } from '../../common/tile';
 import { ConcordanceTileModel } from './model';
 import { init as viewInit } from './views';
