+++ src/js/tiles/concordance/model.ts
@@ -24,7 +24,8 @@ import { concatMap } from 'rxjs/operators';
 import { AppServices } from '../../appServices';
 import { Line, ConcResponse } from '../../common/api/abstract/concordance';
 import { ConcordanceMinState, IStateArgsMapper } from '../../common/models/concordance';
-import { HTTPMethod, SystemMessageType, isSubqueryPayload, DataApi } from '../../common/types';
+import { HTTPMethod, SystemMessageType, DataApi } from '../../common/types';
+import { isSubqueryPayload } from '../../common/query';
 import { Backlink, BacklinkWithArgs } from '../../common/tile';
 import { ActionName as GlobalActionName, Actions as GlobalActions } from '../../models/actions';
 import { findCurrLemmaVariant, WdglanceMainFormModel } from '../../models/query';
