+++ src/js/tiles/empty.ts
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { QueryType } from '../common/types';
+import { QueryType } from '../common/query';
 import { ITileProvider, TileComponent } from '../common/tile';
