+++ src/js/tiles/freqBar/index.ts
@@ -21,7 +21,8 @@ import { ActionDispatcher, ViewUtils } from 'kombo';
 import { AppServices } from '../../appServices';
 import { DataRow, MultiBlockFreqDistribAPI, FreqSort } from '../../common/api/kontext/freqs';
 import { FreqDataBlock, SubqueryModeConf } from '../../common/models/freq';
-import { LocalizedConfMsg, QueryType } from '../../common/types';
+import { LocalizedConfMsg } from '../../common/types';
+import { QueryType } from '../../common/query';
 import { TileComponent, TileConf, TileFactory, Backlink, ITileProvider } from '../../common/tile';
 import { puid } from '../../common/util';
 import { GlobalComponents } from '../../views/global';
