+++ src/js/tiles/syd/index.ts
@@ -18,7 +18,7 @@
 import * as Immutable from 'immutable';
 
 import { AppServices } from '../../appServices';
-import { QueryType } from '../../common/types';
+import { QueryType } from '../../common/query';
 import { ITileProvider, TileComponent, TileConf, TileFactory } from '../../common/tile';
 import { StrippedFreqResponse, SyDAPI } from './api';
 import { SydModel } from './model';
