+++ src/js/tiles/treq/index.ts
@@ -18,7 +18,7 @@
 import * as Immutable from 'immutable';
 
 import { AppServices } from '../../appServices';
-import { QueryType } from '../../common/types';
+import { QueryType } from '../../common/query';
 import { ITileProvider, TileComponent, TileConf, TileFactory } from '../../common/tile';
 import { SearchPackages, TreqAPI, TreqTranslation } from '../../common/api/treq';
 import { TreqModel } from './model';
